--- src/productionEnum.macro.d.ts
@@ -1,3 +1,0 @@
-function productionEnum(value: string): number
-
-export default productionEnum
